Update Bundler version to run tests on Travis
Trying to run the tests for this repo on Travis gives the following
error:
```
Bundler could not find compatible versions for gem "bundler":
  In Gemfile:
    bundler (~> 1.11) ruby
  Current Bundler version:
    bundler (1.7.9)
```

This seems to be because we’re using quite an old version of the
`bundler` Gem.

Also updates Travis config to use a newer Ruby for the same reason.