Make secret include API key + service ID
In research we’ve seen people mix up the service ID and API key because
they’re both 36 character UUIDs. We can’t get rid of the service ID
because it’s used to look up the API key.

Instead, we should change API key to be one long string, which contains
both the service ID, API key and (optionally) the name of the key. For
example:

**API key**
casework_production-8b3aa916-ec82-434e-b0c5-d5d9b371d6a3-dcdc5083-2fee-4fba-8afd-51f3f4bcb7b0

This commit changes the client to make the `secret` argument optional.
If it’s not passed, they we attempt to extract it from the second
argument (service ID) instead.

If a new-style API key is passed with a service ID as well then we just
strip out the parts of the key that aren’t the secret.