Fix integration tests to run against local code
Previously, the 'require' statements at the top of the integration test
file '/bin/test_client.rb' were requiring the installed copy of the
notifications-ruby-client gem and running the tests against the this. If
the gem was uninstalled, the integration tests would fail.

This commit fixes the integration tests so that they are run against the
code in the repo.