Add precompiled letters
Added a new `send_precompiled_letter` method to the client. This accepts
a File object and reference string for the letter, base64 encodes the content
of the PDF file then sends both to the correct endpoint in notifications-api.

The response for precompiled letters is different to the response for
all other notifications since it has fewer properties (it only contains the `id`
and `reference` of the notification. As a result, there's now a new class,
`ResponsePrecompiledLetter` for the precompiled letter responses.