Add method for uploading documents
Added a helper method, 'prepare_upload', which can be used to prepare a
file type object to send over email as a document. This gets used with
the 'send_email' method.

Also added an extra test for precompiled letters to test sending a
StringIO object instead of a File object.