Handle server errors from the api
We already had code to check if the api returns a client error and to
raise a `RequestError` but we were not checking for or handling server
errors. This meant that if someone tries to send a notification and
there is a server error, instead of seeing that error the response from
the client is a Notification object with `nil` for all its properties.

This commit checks is the response from the api is a
`Net::HTTPServerError` and raises a `RequestError` if it is.