RequestError provides message as string
Currently RequestError doesn't conform with Ruby's `Exception` class which
expects message to be a string [1]. When Raven, the Ruby client for Sentry,
processes the exception an error occur because it tries to slice the string
[2].

This changes RequestError from providing message as an array to providing
message as a string. We also remove the `to_s` method as it's no longer needed
since message is returning a string anyway.

[1]: https://ruby-doc.org/core-2.6.3/Exception.html#method-i-message
[2]: https://github.com/getsentry/raven-ruby/blob/263a33b94d9a689d0fc5ba5c16bc92be9538e9fb/lib/raven/event.rb#L88