Bump Ruby version to 2.6
Ruby version 2.4 is no longer supported. Also upgraded some of the
development dependencies and fixed some deprecated syntax in the tests
as a result.

These changes shouldn't have any effect on users of the gem. Changing the
version of Ruby only changes the version of Ruby we use to run the tests, it
doesn't force users of the gem to have Ruby version 2.6 installed.