Add ability to pass options to the http client
We've found out when there is an increased latency with Notify or, more precisely with some of their upstream services (like, in particular Amazon S3 if attaching documents) the default `Net::HTTP` 60 seconds read timeout is not enough and the request is forcibly closed by the client.

This has a series of problems, the main one being that although the client closed the request, it seems like the email gets sent.

If we don't receive a successful response from Notify API we consider this a failed request and re-enqueue the email to be retried automatically a few seconds later, which in turns can again timeout on the client side, but not on the server side, which will manage to send the email.
In the end, multiple, duplicated emails are sent.

Long story short, we want to be able to configure a custom, higher read timeout to minimise where possible these latency issues.

Instead of forcing a higher time out for all users of this gem, I considered it would be better to let this as it is now (60 seconds is the default from Net::HTTP) and add the ability to pass custom timeouts (read/open) as well as other options, to the client.