Add letter_contact_block to Notifications::Client::Template
This adds a new `letter_contact_block` attribute to the
`Notifications::Client::Template` class. `letter_contact_block` can be
`nil` or the contact block of a letter template and will now be returned
as part of the response from the `get_template_by_id`, `get_template_version`
and `get_all_templates` methods.