Add permanent-failure status for letters
If a letter passes our validation but the print provider can not print the letter we should mark it as permanent-failure. An error message for permanent-failure is being added to the UI.
Technical-failure had the wrong message and validation failure will not show the pdf because the letter is in the wrong bucket.

This PR includes:
- add new status type for letters
- align status descriptions for all client documentation
- move status descriptions under the two get message methods
- use the same status description for all client docs
- fix document links