Stop installing gems within the project directory
This was causing a problem when developing locally, since the line

   -v "pwd:/var/project"

in the run_with_docker.sh script was overwriting the packages that
were written to the container image in the bootstrap task. There's
no reason why the packages have to be installed in that directory,
so this removes the flags to use the system defaults.