Switch to environment.sh instead of docker.env
This makes it easier for people on the Notify team to get started,
since the credentials are stored in an 'export X=Y' format - with
the previous 'docker.env' approach, you would have to manually get
rid of the 'export' part. While this means having to add a lot of
'-e ...' options to the run_with_docker.sh script, we make a saving
in CONTRIBUTING.md. Other advantages of this approach:

- It's more similar to how CI works [1] (this is also why we can't
change the credentials file to work with the 'docker.env' approach).

- It's more similar to how our other apps work [2].

[1]: https://github.com/alphagov/notifications-aws/blob/acb88a1ce23ff1a42c8fd3fda7277657d0b9dcfc/concourse/api-clients-pipeline.yml.j2#L249
[2]: https://github.com/alphagov/notifications-template-preview/blob/f9df02478683896a38b66f4f77d50775729905da/scripts/run_with_docker.sh#L4